requirements.txt
Flask
Flask-SQLAlchemy
charm-crypto
pycryptodome

#To install all the dependencies "pip install -r requirements.txt"